refactor: separate deferred.ts
